Move SelectionSet and Into impl to selection module
